Add support for rouge light/dark mode
The latest version of rouge has both a github light theme and a github
dark theme. With this commit, code highlighting will now correctly use
whichever theme is set in the frontmatter of a page or in _config.yml.
This also supports when the theme is set to an auto theme.